workaround for wrong-order Z3 destructor
For some reason, Python is removing references inside the Z3 module
before it runs all the destructors for that module's instances. This
causes some exception logging to be printed out whenever Python
exits. Since this seems to only happen when the affected objects are
globals, this moves the objects to a BackendConcrete instance, which
shouldn't add too much overhead but removes the errors.

It might be worth looking into a "real" fix sometime :)