slightly different fix for replacement frontend issue
disable replacement of replacements by default, when not using complex replacement. This means that the new, inconsistent constraints will not overwrite the old replacement, properly causing a False to be added to the underlying frontend. The advantage of this method is that the original constraint list is kept intact, with the new, 'unsat-causing' constraint there. This makes it possible to clear preconstraints and keep the last path constraint.