simplifer: push Extract down in Concat expressions
If we have an expression like the following:

    Extract(15, 0, Concat(Reverse(Concat(a, b)), c))  with a, b, c 8-bit expressions

the current code cannot simplify this, since both parts of the concat are used.
But if we push the extract down into the end points, we can simplify this to:

    Concat(Extract(7, 0, Reverse(Concat(a,b))), c)
    = Concat(a, c)

Expressions like this example frequently happen using the current implementation
of SimFastMemory, so we really want to simplify this especially in the case that
a, c are constant and b is not.