BackendZ3: Compare new AST and cached ones to avoid collisions.
Thanks to the recent patch that makes claripy less recursive, we are
probably releasing Z3 objects more often than before. A negative
consequence is that our Z3 AST hash, which uses to heavily rely on memory
addresses of Z3 ASTs, now have a much higher chance to collide. This patch
solves this problem in a painful way: Always check if the cached Z3 object
is the same as the one being abstracted, and do not return the cached one
if they are not equal to avoid AST collisions. It will surely impact the
speed, but it is still better than returning incorrect results.

Thanks @basher for reporting another issue that led us to find this bug.