A better solution.
The old fix introduces way too many equality checks, which greatly slows
down simplifications when ASTs are big and deep. Here is a better
solution: We convert `_ast_cache` into an LRUCache and store both
high-level claripy ASTs and the low-level `z3types.Ast` objects. Most
importantly, we increment the Z3 reference counter of the `z3types.Ast`
object so it will not be immediately released. This allows us to leverage
the internal de-duplication mechanism of Z3 (if there is any), which
hopefully will use the same pointer for AST sub-trees if they are the
same. Hence we can use pointer values of low-level Z3 ASTs to key our
`_ast_cache` dict. When a low-level AST is popped out of the `_ast_cache`
dict, we decrement its reference counter inside Z3.