BackendZ3: Bypass integer string conversion limit.
CPython 3.11 introduced integer string conversion length limit (see
https://docs.python.org/3/library/stdtypes.html#integer-string-conversion-length-limitation).
We determined that this security protection does not apply to the threat
model that angr would face, and in fact, causes issues when we deal with
really long strings and integers. Therefore, we bypass the integer
conversion limit in this PR.

We also monkey-patch the Z3 Python binding so that it can accept long
integers in constraints.